ジャムの開催中止ボタンをコンポーネント化
