+++ components/UncancelJamButton.tsx
@@ -1,0 +1,37 @@
+import { useMutation } from '@apollo/client'
+import { NextPage } from 'next'
+import { useState } from 'react'
+import {
+  UNCANCEL_JAM,
+  UncancelJamData,
+} from '../graphql/mutations/uncancel-jam.mutation'
+import { Jam } from 'types'
+import { Session } from 'next-auth'
+
+interface UncancelJamButtonProps {
+  jam: Jam
+  session: Session
+}
+
+const UncancelJamButton: NextPage<UncancelJamButtonProps> = ({ jam }) => {
+  const [message, setMessage] = useState<string>('')
+  const [uncancelJam, uncancelResult] =
+    useMutation<UncancelJamData>(UNCANCEL_JAM)
+  const uncancelError = uncancelResult.error
+  const handleUncancel = async () => {
+    await uncancelJam({ variables: { id: parseInt(jam.id) } })
+    uncancelError
+      ? alert(`Error: ${JSON.stringify(uncancelError)}`)
+      : setMessage('Successfully uncanceled')
+  }
+  if (!jam.canceledAt) return null
+
+  return (
+    <>
+      <button onClick={handleUncancel}>中止取消</button>
+      {message ? <span className="message">{message}</span> : null}
+    </>
+  )
+}
+
+export default UncancelJamButton
